Consolidate duplicate runtime actuals
AbstractDeserializer and AbstractMessage had identical implementations
across all three platforms; move them to commonMain as concrete classes,
dropping expect/actual.

Bytes, Codec, CollectionFactory, and Message were identical between JS
and Native; introduce a nonJvmMain intermediate source set to share them.

Simplify conformance runner build.gradle.kts to minimize diff from main.
