+++ protokt-runtime/src/commonMain/kotlin/protokt/v1/AbstractMessage.kt
@@ -16,6 +16,10 @@
 package protokt.v1
 
 @OnlyForUseByGeneratedProtoCode
-expect abstract class AbstractMessage() : Message {
-    final override fun serialize(): ByteArray
+abstract class AbstractMessage : Message {
+    final override fun serialize(): ByteArray {
+        val writer = codec.writer(serializedSize())
+        serialize(writer)
+        return writer.toByteArray()
+    }
 }
