--- protokt-runtime/src/jsMain/kotlin/protokt/v1/Bytes.kt
@@ -1,34 +1,0 @@
-/*
- * Copyright (c) 2023 Toast, Inc.
- *
- * Licensed under the Apache License, Version 2.0 (the "License");
- * you may not use this file except in compliance with the License.
- * You may obtain a copy of the License at
- * http://www.apache.org/licenses/LICENSE-2.0
- *
- * Unless required by applicable law or agreed to in writing, software
- * distributed under the License is distributed on an "AS IS" BASIS,
- * WITHOUT WARRANTIES OR CONDITIONS OF ANY KIND, either express or implied.
- * See the License for the specific language governing permissions and
- * limitations under the License.
- */
-
-package protokt.v1
-
-actual class Bytes internal actual constructor(value: ByteArray) : AbstractBytes(value) {
-    // Annotation `@JvmStatic` is missing on actual declaration
-    @Suppress("ACTUAL_ANNOTATIONS_NOT_MATCH_EXPECT")
-    actual companion object {
-        actual fun empty() =
-            AbstractBytes.empty()
-
-        actual fun from(bytes: ByteArray) =
-            AbstractBytes.from(bytes)
-
-        actual fun from(message: Message) =
-            AbstractBytes.from(message)
-    }
-}
-
-internal actual fun clone(bytes: ByteArray) =
-    bytes.copyOf()
