+++ protokt-runtime/src/nonJvmMain/kotlin/protokt/v1/Bytes.kt
